Refactor suite2p rule
