+++ photon_mosaic/rules/suite2p_run.py
@@ -4,30 +4,23 @@ from typing import Optional
 
 from suite2p import run_s2p
 
-from photon_mosaic.s2p_options import get_edited_options
-
+from suite2p.default_ops import default_ops
 
 def run_suite2p(
-    tiff_file: str,
     stat_path: str,
-    bin_path: str,
     dataset_folder: Path,
     user_ops_dict: Optional[dict] = None,
 ):
     """
-    This function runs Suite2P on a given TIFF file and saves the
+    This function runs Suite2P on a given dataset folder and saves the
     results in the specified paths. It also handles any exceptions
     that may occur during the process and logs them in an error
     file.
 
     Parameters
     ----------
-    tiff_file : str
-        The path to the TIFF file to be processed.
     stat_path : str
         The path where the Suite2P statistics will be saved.
-    bin_path : str
-        The path where the Suite2P binary data will be saved.
     dataset_folder : str
         The path to the folder containing the dataset.
     user_ops_dict : dict, optional
@@ -47,3 +40,53 @@ def run_suite2p(
         with open(dataset_folder / "error.txt", "a") as f:
             f.write(f"Error: {e}\n")
             f.write(traceback.format_exc())
+
+
+
+
+def get_edited_options(
+    input_path: Path, save_folder: Path, user_ops_dict: Optional[dict] = None
+) -> dict:
+    """This function generates a dictionary of options for Suite2P
+    by loading the default options and then modifying them with
+    user-provided options. The function also sets the required
+    runtime paths for saving the results.
+
+    Parameters
+    ----------
+    input_path : Path
+        The path to the input data folder.
+    save_folder : Path
+        The path to the folder where the results will be saved.
+    user_ops_dict : dict, optional
+        A dictionary containing user-provided options to override
+        the default options. The default is None.
+
+    Returns
+    -------
+    dict
+        A dictionary containing the Suite2P options, including
+        the user-provided options and the required runtime paths.
+
+    Raises
+    ------
+    ValueError
+        If a user-provided option is not valid for Suite2P.
+    """
+
+    ops = default_ops()
+
+    # Override with user-provided subset of keys
+    if user_ops_dict:
+        for key, val in user_ops_dict.items():
+            if key not in ops:
+                raise ValueError(f"Invalid Suite2p option: {key}")
+            ops[key] = val
+
+    # Add required runtime paths
+    ops["save_folder"] = str(save_folder)
+    ops["save_path0"] = str(save_folder)
+    ops["fast_disk"] = str(save_folder.parent)
+    ops["data_path"] = [str(input_path)]
+
+    return ops
