+++ photon_mosaic/workflow/Snakefile
@@ -48,6 +48,7 @@ SAMPLE_PAIRS = [
 ]
 
 include: "preprocessing.smk"
+include: "suite2p.smk"
 
 preproc_targets = [
     f"{processed_data_base}/sub-{i}_{datasets_new_names[i]}/ses-{j}/funcimg/{output_patterns[j]}"
@@ -65,29 +66,3 @@ suite2p_targets = [
 rule all:
     input:
         preproc_targets + suite2p_targets
-
-
-rule suite2p:
-    input:
-        tiff=lambda wildcards: str(
-            processed_data_base /
-            f"sub-{wildcards.sub_idx}_{datasets_new_names[int(wildcards.sub_idx)]}" /
-            f"ses-{wildcards.ses_idx}/funcimg/" /
-            f"{output_patterns[int(wildcards.ses_idx)]}"
-        )
-    output:
-        F=f"{processed_data_base}/sub-{{sub_idx}}_{{dataset}}/ses-{{ses_idx}}/funcimg/suite2p/plane0/F.npy",
-        bin=f"{processed_data_base}/sub-{{sub_idx}}_{{dataset}}/ses-{{ses_idx}}/funcimg/suite2p/plane0/data.bin"
-    params:
-        dataset_folder=f"{processed_data_base}/sub-{{sub_idx}}_{{dataset}}/ses-{{ses_idx}}/funcimg"
-    resources:
-        **(slurm_config if config.get("use_slurm") else {}),
-    run:
-        from photon_mosaic.rules.suite2p_run import run_suite2p
-        run_suite2p(
-            input.tiff,
-            output.F,
-            output.bin,
-            Path(params.dataset_folder),
-            config["suite2p_ops"],
-        )
